// apps/admin-web/src/.generated/api/model/membershipItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

export interface MembershipItemDto {
  id: string;
  code: string;
  /** @nullable */
  label: string | null;
  /** @nullable */
  description: string | null;
  isSystem: boolean;
  customerCount: number;
  permissions: string[];
}